Extreme streamlining reduce basic tests and renamed from 'sanity' to 'basic'
