+++ test/functional/forall/reduce-basic/tests/test-forall-basic-ReduceMaxLoc.hpp
@@ -5,8 +5,8 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
-#ifndef __TEST_FORALL_REDUCEMINLOC_SANITY_HPP__
-#define __TEST_FORALL_REDUCEMINLOC_SANITY_HPP__
+#ifndef __TEST_FORALL_BASIC_REDUCEMAXLOC_HPP__
+#define __TEST_FORALL_BASIC_REDUCEMAXLOC_HPP__
 
 #include <cstdlib>
 #include <numeric>
@@ -14,7 +14,7 @@
 
 template <typename DATA_TYPE, typename WORKING_RES, 
           typename EXEC_POLICY, typename REDUCE_POLICY>
-void ForallReduceMinLocSanityTestImpl(RAJA::Index_type first, RAJA::Index_type last)
+void ForallReduceMaxLocBasicTestImpl(RAJA::Index_type first, RAJA::Index_type last)
 {
   RAJA::TypedRangeSegment<RAJA::Index_type> r1(first, last);
 
@@ -30,60 +30,60 @@ void ForallReduceMinLocSanityTestImpl(RAJA::Index_type first, RAJA::Index_type l
                                     &test_array);
 
   const int modval = 100;
-  const DATA_TYPE min_init = modval+1;
-  const RAJA::Index_type minloc_init = -1;
-  const RAJA::Index_type minloc_idx = (last - first) * 2/3 + first;
-  const DATA_TYPE small_min = -modval;
-  const RAJA::Index_type small_minloc = minloc_init;
+  const DATA_TYPE max_init = -modval;
+  const RAJA::Index_type maxloc_init = -1;
+  const RAJA::Index_type maxloc_idx = (last - first) * 2/3 + first;
+  const DATA_TYPE big_max = modval+1;
+  const RAJA::Index_type big_maxloc = maxloc_init;
 
   for (RAJA::Index_type i = 0; i < last; ++i) {
     test_array[i] = static_cast<DATA_TYPE>( rand() % modval );
   }
-  test_array[minloc_idx] = static_cast<DATA_TYPE>(small_min);
+  test_array[maxloc_idx] = static_cast<DATA_TYPE>(big_max);
 
-  DATA_TYPE ref_min = min_init;
-  RAJA::Index_type ref_minloc = minloc_init;
+  DATA_TYPE ref_max = max_init;
+  RAJA::Index_type ref_maxloc = maxloc_init;
   for (RAJA::Index_type i = first; i < last; ++i) {
-    if ( test_array[i] < ref_min ) {
-       ref_min = test_array[i];
-       ref_minloc = i;
+    if ( test_array[i] > ref_max ) {
+       ref_max = test_array[i];
+       ref_maxloc = i;
     } 
   }
 
   working_res.memcpy(working_array, test_array, sizeof(DATA_TYPE) * last);
 
 
-  RAJA::ReduceMinLoc<REDUCE_POLICY, DATA_TYPE, RAJA::Index_type> mininit(small_min, minloc_init);
-  RAJA::ReduceMinLoc<REDUCE_POLICY, DATA_TYPE, RAJA::Index_type> min(min_init, minloc_init);
+  RAJA::ReduceMaxLoc<REDUCE_POLICY, DATA_TYPE, RAJA::Index_type> maxinit(big_max, maxloc_init);
+  RAJA::ReduceMaxLoc<REDUCE_POLICY, DATA_TYPE, RAJA::Index_type> max(max_init, maxloc_init);
 
   RAJA::forall<EXEC_POLICY>(r1, [=] RAJA_HOST_DEVICE(RAJA::Index_type idx) {
-    mininit.minloc( working_array[idx], idx );
-    min.minloc( working_array[idx], idx );
+    maxinit.maxloc( working_array[idx], idx );
+    max.maxloc( working_array[idx], idx );
   });
 
-  ASSERT_EQ(static_cast<DATA_TYPE>(mininit.get()), small_min);
-  ASSERT_EQ(static_cast<RAJA::Index_type>(mininit.getLoc()), small_minloc);
-  ASSERT_EQ(static_cast<DATA_TYPE>(min.get()), ref_min);
-  ASSERT_EQ(static_cast<RAJA::Index_type>(min.getLoc()), ref_minloc);
+  ASSERT_EQ(static_cast<DATA_TYPE>(maxinit.get()), big_max);
+  ASSERT_EQ(static_cast<RAJA::Index_type>(maxinit.getLoc()), big_maxloc);
+  ASSERT_EQ(static_cast<DATA_TYPE>(max.get()), ref_max);
+  ASSERT_EQ(static_cast<RAJA::Index_type>(max.getLoc()), ref_maxloc);
 
-  min.reset(min_init, minloc_init);
-  ASSERT_EQ(static_cast<DATA_TYPE>(min.get()), min_init);
-  ASSERT_EQ(static_cast<RAJA::Index_type>(min.getLoc()), minloc_init);
+  max.reset(max_init, maxloc_init);
+  ASSERT_EQ(static_cast<DATA_TYPE>(max.get()), max_init);
+  ASSERT_EQ(static_cast<RAJA::Index_type>(max.getLoc()), maxloc_init);
 
   DATA_TYPE factor = 2;
   RAJA::forall<EXEC_POLICY>(r1, [=] RAJA_HOST_DEVICE(RAJA::Index_type idx) {
-    min.minloc( working_array[idx] * factor, idx);
+    max.maxloc( working_array[idx] * factor, idx);
   });
-  ASSERT_EQ(static_cast<DATA_TYPE>(min.get()), ref_min * factor);
-  ASSERT_EQ(static_cast<RAJA::Index_type>(min.getLoc()), ref_minloc);
-
+  ASSERT_EQ(static_cast<DATA_TYPE>(max.get()), ref_max * factor);
+  ASSERT_EQ(static_cast<RAJA::Index_type>(max.getLoc()), ref_maxloc);
+  
   factor = 3;
   RAJA::forall<EXEC_POLICY>(r1, [=] RAJA_HOST_DEVICE(RAJA::Index_type idx) { 
-    min.minloc( working_array[idx] * factor, idx);
+    max.maxloc( working_array[idx] * factor, idx);
   });
-  ASSERT_EQ(static_cast<DATA_TYPE>(min.get()), ref_min * factor);
-  ASSERT_EQ(static_cast<RAJA::Index_type>(min.getLoc()), ref_minloc);
-   
+  ASSERT_EQ(static_cast<DATA_TYPE>(max.get()), ref_max * factor);
+  ASSERT_EQ(static_cast<RAJA::Index_type>(max.getLoc()), ref_maxloc);
+ 
 
   deallocateForallTestData<DATA_TYPE>(working_res,
                                       working_array,
@@ -91,28 +91,28 @@ void ForallReduceMinLocSanityTestImpl(RAJA::Index_type first, RAJA::Index_type l
                                       test_array);
 }
 
-TYPED_TEST_SUITE_P(ForallReduceMinLocSanityTest);
+TYPED_TEST_SUITE_P(ForallReduceMaxLocBasicTest);
 template <typename T>
-class ForallReduceMinLocSanityTest : public ::testing::Test
+class ForallReduceMaxLocBasicTest : public ::testing::Test
 {
 };
 
-TYPED_TEST_P(ForallReduceMinLocSanityTest, ReduceMinLocSanityForall)
+TYPED_TEST_P(ForallReduceMaxLocBasicTest, ReduceMaxLocBasicForall)
 {
   using DATA_TYPE     = typename camp::at<TypeParam, camp::num<0>>::type;
   using WORKING_RES   = typename camp::at<TypeParam, camp::num<1>>::type;
   using EXEC_POLICY   = typename camp::at<TypeParam, camp::num<2>>::type;
   using REDUCE_POLICY = typename camp::at<TypeParam, camp::num<3>>::type;
 
-  ForallReduceMinLocSanityTestImpl<DATA_TYPE, WORKING_RES, 
+  ForallReduceMaxLocBasicTestImpl<DATA_TYPE, WORKING_RES, 
                                    EXEC_POLICY, REDUCE_POLICY>(0, 28);
-  ForallReduceMinLocSanityTestImpl<DATA_TYPE, WORKING_RES, 
+  ForallReduceMaxLocBasicTestImpl<DATA_TYPE, WORKING_RES, 
                                    EXEC_POLICY, REDUCE_POLICY>(3, 642);
-  ForallReduceMinLocSanityTestImpl<DATA_TYPE, WORKING_RES, 
+  ForallReduceMaxLocBasicTestImpl<DATA_TYPE, WORKING_RES, 
                                    EXEC_POLICY, REDUCE_POLICY>(0, 2057);
 }
 
-REGISTER_TYPED_TEST_SUITE_P(ForallReduceMinLocSanityTest,
-                            ReduceMinLocSanityForall);
+REGISTER_TYPED_TEST_SUITE_P(ForallReduceMaxLocBasicTest,
+                            ReduceMaxLocBasicForall);
 
-#endif  // __TEST_FORALL_REDUCEMINLOC_SANITY_HPP__
+#endif  // __TEST_FORALL_BASIC_REDUCEMAXLOC_HPP__
